Move TS print into the same file
