+++ src/language-js/print/mapped-type.js
@@ -1,6 +1,13 @@
+import { printDanglingComments } from "../../main/comments/print.js";
 import hasNewlineInRange from "../../utils/has-newline-in-range.js";
 import { locStart, locEnd } from "../loc.js";
-import { group, softline, indent, ifBreak } from "../../document/builders.js";
+import {
+  group,
+  softline,
+  indent,
+  ifBreak,
+  line,
+} from "../../document/builders.js";
 
 /**
  * @param {string | null} optional
@@ -21,7 +28,6 @@ function printFlowMappedTypeOptionalModifier(optional) {
 
 function printFlowMappedTypeProperty(path, options, print) {
   const { node } = path;
-
   const shouldBreak = hasNewlineInRange(
     options.originalText,
     locStart(node),
@@ -48,4 +54,50 @@ function printFlowMappedTypeProperty(path, options, print) {
   );
 }
 
-export { printFlowMappedTypeProperty };
+/**
+ * @param {string} tokenNode
+ * @param {string} keyword
+ * @returns {string}
+ */
+function printTypeScriptMappedTypeModifier(tokenNode, keyword) {
+  if (tokenNode === "+" || tokenNode === "-") {
+    return tokenNode + keyword;
+  }
+
+  return keyword;
+}
+
+function printTypescriptMappedType(path, options, print) {
+  const { node } = path;
+  const shouldBreak = hasNewlineInRange(
+    options.originalText,
+    locStart(node),
+    locEnd(node)
+  );
+
+  return group(
+    [
+      "{",
+      indent([
+        options.bracketSpacing ? line : softline,
+        print("typeParameter"),
+        node.optional
+          ? printTypeScriptMappedTypeModifier(node.optional, "?")
+          : "",
+        node.typeAnnotation ? ": " : "",
+        print("typeAnnotation"),
+        options.semi ? ifBreak(";") : "",
+      ]),
+      printDanglingComments(path, options),
+      options.bracketSpacing ? line : softline,
+      "}",
+    ],
+    { shouldBreak }
+  );
+}
+
+export {
+  printFlowMappedTypeProperty,
+  printTypeScriptMappedTypeModifier,
+  printTypescriptMappedType,
+};
