+++ src/language-js/utils/index.js
@@ -486,19 +486,6 @@ function isSimpleTemplateLiteral(node) {
 }
 
 /**
- * @param {string} tokenNode
- * @param {string} keyword
- * @returns {string}
- */
-function getTypeScriptMappedTypeModifier(tokenNode, keyword) {
-  if (tokenNode === "+" || tokenNode === "-") {
-    return tokenNode + keyword;
-  }
-
-  return keyword;
-}
-
-/**
  * @param {string | null} optional
  * @returns {string}
  */
@@ -1149,7 +1136,6 @@ export {
   getFlowMappedTypeOptionalModifier,
   getLeftSide,
   getLeftSidePathName,
-  getTypeScriptMappedTypeModifier,
   hasLeadingOwnLineComment,
   hasNakedLeftSide,
   hasNode,
